Enable G-Key macro: Sending definitions is working
This is the second development step for treating
the Macro keys G1-18 on the K95 RGB keyboard.

If you type a valid key sequence in the
Macro Tab -> Macro Key Actions,
select at least one G-Key and click Apply,
the macro definiton ist sent
to the keyboard and active.
The Macro definition is saved like all other bindings,
so with restart of ckb the macro definitions
are re-sent to the keyboard,
but you can't see the older definitions
in the ckb-panel yet.

Reset and Unbind are working properly,
Reset to Default is not tested in that commit.

Some debug outputs are still active (qDebug)
and some ToDos are marked in the source.