setup ckb for client side macro managment (aplha)
This branch sets up the ckb client (and one addition in ckb-daemon)
for self defined macros at nearly each key. With this enhancement
the 18 G-Keys at the K95 can be used.

It adds a new tab "macro" to the bind page as user interface
and includes all changes and additions to run the new tab
or do the communication with the daemon.

This version 0.0.1 handles the key macros in their primary layer,
so shift, alt, alt-Gr or meta keys are ignored.
It it tested with a K95 RGB on linux (mint 17.1 and 17.2) on
different machines.