Fix initialization bug with K70RGB + M65RGB
For the combination of an K70RGB keyboard and the M65RGB mouse
there was an initialization bug.
The ckb client did send a "Macro clear" command to the driver
for the mouse. Because that is unneccessary, there ist no function
to handle this and the deamon got a SIGSEGV.

This patch fixes that bug.
The change in ckb client prevents issuing the command.
The Patch in ckb-daemon may be activated by setting INIT_ERROR
and only does some debug output.